backend/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.22.0
sqlmodel==0.0.9
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-dotenv==1.0.0
httpx==0.24.0
stripe==6.0.0
pytest==7.4.4
